# lib/api-zod/src/generated/types/aiReviewBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeSync - Collaborative Web IDE API
 * OpenAPI spec version: 0.1.0
 */

export interface AiReviewBody {
  code: string;
  language: string;
  roomId?: string;
  fileId?: string;
}