# lib/api-zod/src/generated/types/aiReviewResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeSync - Collaborative Web IDE API
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewIssue } from "./reviewIssue";

export interface AiReviewResult {
  issues: ReviewIssue[];
  summary: string;
  suggestions: string[];
}